start slope migration - move code to slope.py
